Also sign the kernel image itself for use with Secure Boot
This will NOT work with UKIs as they are always generated locally and therefore
have to be signed locally. USE=secureboot on gentoo-kernel-bin accomplishes this
but does require the user to have their own key.

Actually booting with Secure Boot enabled will require that the user adds
our certificate to their UEFI key database or to their shim MOK list.

I suggest to just use the same key as for the modules to keep things simple.

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>
Closes: https://github.com/projg2/binpkg-docker/pull/3
Signed-off-by: Michał Górny <mgorny@gentoo.org>