*: move installkernel-{gentoo,systemd} -> installkernel
We disable USE=initramfs on the kernel to avoid a dependency conflict
where kernel-install.eclass wants +dracut on arches where we do not build
generic-uki, but -dracut on arches where we do build a generic-uki.

It doesn't really matter for us that no initramfs will be generated in postinst
anyway, in fact it might make the builds ever so slightly faster.

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>
Closes: https://github.com/projg2/binpkg-docker/pull/6
Signed-off-by: Michał Górny <mgorny@gentoo.org>